JA-001 | reduce module -> class redundancy using a single module instead of a package
